requirement.txt
asgiref==3.2.7
astroid==2.3.3
autopep8==1.5
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
Django==3.0.4
django-allauth==0.41.0
django-crispy-forms==1.9.0
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
Pillow==7.0.0
pycodestyle==2.5.0
pylint==2.4.4
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.1
urllib3==1.25.8
wrapt==1.11.2
